Inline url_encoding.ts
